Demo's/1/pyproject.toml
[project]
name = "1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=26.3.1",
    "isort>=8.0.1",
    "langchain>=1.2.15",
    "langchain-google-genai>=4.2.1",
    "python-dotenv>=1.2.2",
]
